[PRTS-2189] Partnership for Capital (#35380)
feat(prts_for_capital): adds experiments for partnership for capital
feat(prts_for_capital): adds capital product constant
feat(prts_for_capital): adds partner referral link creation for capital
feat(prts_for_capital): adds support for fetching and filtering capital
feat(prts_for_capital): adds support for fetching sub-merchant by ID and product 
feat(prts_for_capital): allows a partner to make an LOS request for a sub-merchant 
feat(prts_for_capital): create CC application, tags & attributes for capital sub-merchant
feat(prts_for_capital): create application after saving details
feat(prts_for_capital): refactor features/service 
feat(prts_for_capital): adds test cases for fetching capital sub-merchant 
feat(prts_for_capital): refactors based on PR reviews

+++ app/Jobs/SubMerchantTaggingJob.php
@@ -2,10 +2,11 @@
 
 namespace RZP\Jobs;
 
+use Throwable;
+use Monolog\Logger;
 use RZP\Trace\TraceCode;
 use RZP\Models\Merchant;
 use RZP\Models\Merchant\Metric;
-use Razorpay\Trace\Logger as Trace;
 use Jitendra\Lqext\TransactionAware;
 
 class SubMerchantTaggingJob extends Job
@@ -27,11 +28,14 @@ class SubMerchantTaggingJob extends Job
 
     protected $subMerchantId;
 
-    public function __construct($mode, string $partnerId, string $subMerchantId)
+    protected $tagPrefix;
+
+    public function __construct($mode, string $partnerId, string $subMerchantId, string $tagPrefix)
     {
         parent::__construct($mode);
         $this->subMerchantId = $subMerchantId;
         $this->partnerId     = $partnerId;
+        $this->tagPrefix     = $tagPrefix;
     }
 
     public function handle()
@@ -43,6 +47,7 @@ class SubMerchantTaggingJob extends Job
             [
                 'partner_id'  => $this->partnerId,
                 'merchant_id' => $this->subMerchantId,
+                'tag_prefix'  => $this->tagPrefix,
             ]
         );
 
@@ -50,18 +55,30 @@ class SubMerchantTaggingJob extends Job
         {
             $partner     = $this->repoManager->merchant->findOrFailPublic($this->partnerId);
             $subMerchant = $this->repoManager->merchant->findOrFailPublic($this->subMerchantId);
-            (new Merchant\Core())->addSubMerchantReferral($partner, $subMerchant);
+
+            $existingTags = $subMerchant->tagNames();
+
+            $refTag = $this->tagPrefix . $partner->getId();
+
+            if (in_array(strtolower($refTag), array_map('strtolower', $existingTags)) === true)
+            {
+                return;
+            }
+
+            (new Merchant\Core())->appendTag($subMerchant, $refTag);
+
             $this->delete();
         }
-        catch (\Throwable $e)
+        catch (Throwable $e)
         {
             $this->trace->traceException(
                 $e,
-                Trace::ERROR,
+                Logger::ERROR,
                 TraceCode::SUBMERCHANT_TAGGING_ASYNC_JOB_FAILED,
                 [
                     'partner_id'  => $this->partnerId,
                     'merchant_id' => $this->subMerchantId,
+                    'tag_prefix'  => $this->tagPrefix
                 ]
             );
 
@@ -69,18 +86,21 @@ class SubMerchantTaggingJob extends Job
         }
     }
 
-    protected function checkRetry(\Throwable $e)
+    protected function checkRetry(Throwable $e)
     {
         $this->countJobException($e);
 
         if ($this->attempts() > self::MAX_RETRY_ATTEMPT)
         {
-            $this->trace->error(TraceCode::SUBMERCHANT_TAGGING_ASYNC_JOB_MESSAGE_DELETE, [
-                'partner_id'   => $this->partnerId,
-                'merchant_id'  => $this->subMerchantId,
-                'job_attempts' => $this->attempts(),
-                'message'      => 'Deleting the job after configured number of tries. Still unsuccessful.'
-            ]);
+            $this->trace->error(
+                TraceCode::SUBMERCHANT_TAGGING_ASYNC_JOB_MESSAGE_DELETE,
+                [
+                    'partner_id'   => $this->partnerId,
+                    'merchant_id'  => $this->subMerchantId,
+                    'job_attempts' => $this->attempts(),
+                    'message'      => 'Deleting the job after configured number of tries. Still unsuccessful.'
+                ]
+            );
 
             $this->trace->count(Metric::SUBMERCHANT_TAGGING_FAILURE_TOTAL, []);
